Use flag to indicate need to rerender
Because we now defer refresh, the refreshes and thus calls to render_plot() are now batched when there is a large number of events in the queue. This happens when you're zooming in or out with kinetic scrolling. The first call sets the needs_rerender flag to False, and subsequent calls will do nothing, until something _has_ actually changed. This makes the widget much more responsive.